requirements.txt
Flask
tensorflow
numpy
pandas
openpyxl
scikit-learn
Pillow
matplotlib
requests
Werkzeug
